accelerate tests
